// src/api/model/deadline-type.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Vector
 * The Vector API
 * OpenAPI spec version: 1.0
 */

export type DeadlineType = (typeof DeadlineType)[keyof typeof DeadlineType]

export const DeadlineType = {
	IMMEDIATE: "IMMEDIATE",
	DATE: "DATE",
	ROLLING: "ROLLING",
} as const